added joining and leaving clubs
